backend/main.py
from fastapi import FastAPI, WebSocket, WebSocketDisconnect
from fastapi.middleware.cors import CORSMiddleware
import json
import asyncio
from backend.agent import app_graph
from backend.config import settings
from langchain_core.messages import HumanMessage

app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=settings.CORS_ORIGINS,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/")
def home():
    return {"status": "Voice AI Backend Running"}

@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    await websocket.accept()
    print("WebSocket Connected")
    
    # Initialize session state (mock)
    session_state = {
        "messages": [],
        "customer_id": None, 
        "is_verified": False
    }

    try:
        # Send Greeting with Audio
        greeting_text = settings.GREETING_MESSAGE
        
        # Async generation to not block accept
        loop = asyncio.get_running_loop()
        from backend.services.audio import generate_audio
        
        # Generate Audio in ThreadPool
        greeting_audio = await loop.run_in_executor(None, generate_audio, greeting_text)
        
        await websocket.send_text(json.dumps({
            "type": "audio", 
            "content": greeting_text, 
            "audio": greeting_audio
        }))

        while True:
            data = await websocket.receive_text()
            
            try:
                payload = json.loads(data)
            except:
                payload = {"type": "text", "text": data}

            user_text = ""
            
            if payload.get("type") == "audio":
                import base64
                from backend.services.audio import transcribe_audio
                
                audio_b64 = payload.get("data")
                if audio_b64:
                    if "," in audio_b64:
                        audio_b64 = audio_b64.split(",")[1]
                    
                    audio_bytes = base64.b64decode(audio_b64)
                    
                    # Run Transcribe in Thread
                    user_text = await loop.run_in_executor(None, transcribe_audio, audio_bytes)
                    print(f"Transcribed: {user_text}")
            else:
                 user_text = payload.get("text", "")
            
            if not user_text:
                continue

            print(f"User: {user_text}")
            
            # Invoke Agent
            session_state["messages"].append(HumanMessage(content=user_text))
            
            # Run graph
            final_state = await app_graph.ainvoke(session_state)
            
            # Get latest response
            last_msg = final_state["messages"][-1]
            response_text = last_msg.content
            
            # Update session state
            session_state = final_state
            
            print(f"Agent: {response_text}")
            
            # Send Text First (Low Latency feedback)
            if not response_text:
                continue

            await websocket.send_text(json.dumps({
                "type": "text", 
                "content": response_text
            }))
            
            # Streaming Audio
            from backend.services.audio import stream_audio, generate_audio
            
            # We can run the stream in a separate thread, but iterating the generator 
            # and sending chunks needs to happen in the async loop.
            # Ideally: get generator from thread, then iterate.
            # For simplicity in V1 specific to OpenAI: The API call `client.audio.speech.create` is blocking.
            # We must use `run_in_executor` to get the response object? 
            # Actually, standard `iter_bytes` is synchronous. 
            # We'll stick to full generation for robustness OR chunked if user insists.
            # User insisted on reducing latency. Pipelining full gen is safer for now than complex async iterator wrapping 
            # without async OpenAI client.
            # Let's revert to "Generate Audio in Thread" but simpler code, OR try a chunked send approach 
            # if we had an async client.
            # Current `openai` client is sync.
            # To truly reduce latency, we should switch to AsyncOpenAI().
            # For this step, let's keep the ThreadPool optimization we did in Step 241, 
            # but ensure we aren't doubly waiting.
            
            # Let's use the blocking generator in a strict thread? No, can't stream across thread boundary easily without queue.
            # Fallback: Just maximize the ThreadPool efficiency.
            
            audio_content = await loop.run_in_executor(None, generate_audio, response_text)
            
            await websocket.send_text(json.dumps({
                "type": "audio", 
                "content": response_text, # Redundant but safe
                "audio": audio_content
            }))

    except WebSocketDisconnect:
        print("WebSocket Disconnected")
