backend/requirements.txt
fastapi
uvicorn
websockets
langgraph
langchain
pydantic
python-dotenv
httpx
langchain-openai
openai
